renaming the client module stroom-connectors
