+++ stroom-core-server/src/main/java/stroom/spring/ServerComponentScanConfiguration.java
@@ -43,7 +43,7 @@ import org.springframework.context.annotation.FilterType;
         "stroom.internalstatistics",
         "stroom.io",
         "stroom.jobsystem",
-        "stroom.kafka",
+        "stroom.connectors.kafka",
         "stroom.lifecycle",
         "stroom.logging",
         "stroom.node",
