+++ stroom-integrationtest/src/test/java/stroom/spring/ProcessTestServerComponentScanConfiguration.java
@@ -92,7 +92,7 @@ import stroom.xmlschema.server.XMLSchemaServiceImpl;
         "stroom.internalstatistics",
         "stroom.io",
         "stroom.jobsystem",
-        "stroom.kafka",
+        "stroom.connectors.kafka",
         "stroom.lifecycle",
         "stroom.logging",
         "stroom.node",
