+++ stroom-integrationtest/src/test/java/stroom/test/SetupSampleDataComponentScanConfiguration.java
@@ -95,7 +95,7 @@ import stroom.xmlschema.server.MockXMLSchemaService;
         "stroom.internalstatistics",
         "stroom.io",
         "stroom.jobsystem",
-        "stroom.kafka",
+        "stroom.connectors.kafka",
         "stroom.lifecycle",
         "stroom.logging",
         "stroom.node",
