+++ stroom-kafka-client-impl_0_10_0_1/build.gradle
@@ -1,24 +1,8 @@
-group 'stroom'
-version 'SNAPSHOT'
-
-apply plugin: 'java'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
 dependencies {
     compile project(':stroom-util')
-    compile project(':stroom-kafka-client')
+    compile project(':stroom-connectors')
 
     compile group: 'org.apache.kafka', name:'kafka-clients', version:'0.10.0.1' //CDH5.10 uses kafka 10.0, Kafka <10.2 is picky about client and server versions.
-
-    compile urlDependencies.get(urlLibs.stroomProxyRepo)
-    compile urlDependencies.get(urlLibs.eventLogging)
-    compile urlDependencies.get(urlLibs.stroomQueryCommon)
-    compile urlDependencies.get(urlLibs.stroomStatsApi)
 
     compile libs.guava
     compile libs.slf4j_api
