+++ stroom-kafka-client-impl_0_10_0_1/src/main/java/stroom/connectors/kafka/StroomKafkaProducerImpl_0_10_0_1.java
@@ -1,4 +1,4 @@
-package stroom.kafka;
+package stroom.connectors.kafka;
 
 import com.google.common.base.Strings;
 import org.apache.kafka.clients.producer.KafkaProducer;
