--- stroom-kafka-client/build.gradle
@@ -1,13 +1,0 @@
-group 'stroom'
-version 'SNAPSHOT'
-
-apply plugin: 'java'
-
-sourceCompatibility = 1.8
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-}
