+++ stroom-pipeline/build.gradle
@@ -7,7 +7,7 @@ dependencies {
     permitUnusedDeclared project(':stroom-statistics-api') //not clear why it thinks this is unused
     compile project(':stroom-util')
     compile project(':stroom-util-shared')
-    compile project(':stroom-kafka-client')
+    compile project(':stroom-connectors')
 
     compile urlDependencies.get(urlLibs.hadoopCommonShaded)
     compile urlDependencies.get(urlLibs.hadoopHdfsShaded)
