+++ stroom-statistics-server/src/main/java/stroom/statistics/server/stroomstats/internal/StroomStatsInternalStatisticsService.java
@@ -5,9 +5,9 @@ import com.google.common.base.Strings;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Component;
-import stroom.kafka.FlushMode;
-import stroom.kafka.StroomKafkaProducer;
-import stroom.kafka.StroomKafkaProducerRecord;
+import stroom.connectors.kafka.FlushMode;
+import stroom.connectors.kafka.StroomKafkaProducer;
+import stroom.connectors.kafka.StroomKafkaProducerRecord;
 import stroom.node.server.StroomPropertyService;
 import stroom.query.api.v2.DocRef;
 import stroom.statistics.internal.InternalStatisticEvent;
