+++ stroom-statistics-server/src/test/java/stroom/statistics/server/stroomstats/internal/TestStroomStatsInternalStatisticsService.java
@@ -10,7 +10,7 @@ import org.mockito.Mockito;
 import org.mockito.runners.MockitoJUnitRunner;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import stroom.kafka.StroomKafkaProducer;
+import stroom.connectors.kafka.StroomKafkaProducer;
 import stroom.node.server.MockStroomPropertyService;
 import stroom.query.api.v2.DocRef;
 import stroom.statistics.internal.InternalStatisticEvent;
